+++ ts4uc/agents/ppo/train.py
@@ -430,7 +430,7 @@ if __name__ == "__main__":
 	parser.add_argument('--update_epochs', type=float, required=False, default=4)
 	parser.add_argument('--clip_ratio', type=float, required=False, default=0.1)
 	parser.add_argument('--forecast_horizon_hrs', type=int, required=False, default=12)
-	parser.add_argument('--credit_assignment_1hr', type=int, required=False, default=0.9)
+	parser.add_argument('--credit_assignment_1hr', type=float, required=False, default=0.9)
 	parser.add_argument('--observation_processor', type=str, required=False, default='LimitedHorizonProcessor')
